moved plugins to the plugin folder
