+++ unity-renderer/Assets/DCLPlugins/CameraReelPlugin.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 61182dff26744dc39bd48cad3e7a527d
+timeCreated: 1690837975
