+++ unity-renderer/Assets/DCLPlugins/CameraReelPlugin/CameraReelPlugin.asmdef
@@ -1,0 +1,23 @@
+{
+    "name": "CameraReelPlugin",
+    "rootNamespace": "",
+    "references": [
+        "GUID:fbcc413e192ef9048811d47ab0aca0c0",
+        "GUID:a3ceb534947fac14da25035bc5c24788",
+        "GUID:fb3a66902b74ab843809d0c30d65f92e",
+        "GUID:7fe146b43e76fe745aeaf2020feb54b7",
+        "GUID:28f74c468a54948bfa9f625c5d428f56",
+        "GUID:1e6b57fe78f7b724e9567f29f6a40c2c",
+        "GUID:f51ebe6a0ceec4240a699833d6309b23",
+        "GUID:3b80b0b562b1cbc489513f09fc1b8f69"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
