+++ unity-renderer/Assets/DCLPlugins/CameraReelPlugin/CameraReelPlugin.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 66123f834d0e1bf46934a130bcd70437
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
