+++ unity-renderer/Assets/Features/CameraReel/ScreenshotViewer/Scripts/ScreenshotViewerController.cs
@@ -1,6 +1,7 @@
 ﻿using Cysharp.Threading.Tasks;
 using DCL;
 using DCLServices.CameraReelService;
+using Features.CameraReel.Section;
 using System;
 using System.Threading.Tasks;
 using UnityEngine;
