+++ unity-renderer/Assets/Features/CameraReel/Section/Prefabs/CameraReelSectionView.prefab
