+++ unity-renderer/Assets/Features/CameraReel/Section/Scripts/CameraReelModel.cs.meta
