+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/PluginSystem/PluginSystem.asmdef
@@ -71,7 +71,8 @@
         "GUID:ab088438f3b94a0da2e753ee6cae5e25",
         "GUID:b2d6f791b7364903976be461bae7f03d",
         "GUID:ad49836f54104ebe85262b0f0f042335",
-        "GUID:fb3a66902b74ab843809d0c30d65f92e"
+        "GUID:fb3a66902b74ab843809d0c30d65f92e",
+        "GUID:66123f834d0e1bf46934a130bcd70437"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
